+++ test/unit/sablier-v2-linear/withdraw-all-to/withdrawAllTo.t.sol
@@ -486,7 +486,7 @@ contract SablierV2Linear__WithdrawAllTo is SablierV2LinearUnitTest {
     {
         // Create the ended dai stream.
         changePrank(daiStream.sender);
-        uint256 earlyStopTime = daiStream.startTime + TIME_OFFSET;
+        uint64 earlyStopTime = daiStream.startTime + TIME_OFFSET;
         uint256 endedDaiStreamId = sablierV2Linear.create(
             daiStream.sender,
             users.recipient,
@@ -537,7 +537,7 @@ contract SablierV2Linear__WithdrawAllTo is SablierV2LinearUnitTest {
     {
         // Create the ended dai stream.
         changePrank(daiStream.sender);
-        uint256 earlyStopTime = daiStream.startTime + TIME_OFFSET;
+        uint64 earlyStopTime = daiStream.startTime + TIME_OFFSET;
         uint256 endedDaiStreamId = sablierV2Linear.create(
             daiStream.sender,
             users.recipient,
